Change services

+++ src/main/java/com/example/springseaport/service/CargoService.java
@@ -2,37 +2,40 @@ package com.example.springseaport.service;
 
 import com.example.springseaport.dto.CargoDto;
 import com.example.springseaport.entity.Cargo;
+import com.example.springseaport.mapper.CargoMapper;
 import com.example.springseaport.repository.CargoRepository;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
+import java.util.stream.Collectors;
 
 @Service
 public class CargoService {
 
     private final CargoRepository cargoRepository;
+    private final CargoMapper mapper;
 
-    public CargoService(CargoRepository cargoRepository) {
+    public CargoService(CargoRepository cargoRepository, CargoMapper cargoMapper) {
         this.cargoRepository = cargoRepository;
+        this.mapper = cargoMapper;
     }
 
     public List<CargoDto> getAllDto() {
-        List<Cargo> cargos = cargoRepository.getAll();
-        List<CargoDto> cargoDtos = new ArrayList<>();
-        for (Cargo cargo : cargos) {
-            cargoDtos.add(new CargoDto(cargo.getId(), cargo.getName(), cargo.getType(), cargo.getWeight()));
-        }
-        return cargoDtos;
+        return cargoRepository.findAll().stream().map(mapper::toDto).collect(Collectors.toList());
     }
 
     public void addCargo(CargoDto cargoDto){
-        cargoRepository.add(new Cargo(cargoRepository.count() + 1, cargoDto.getName(), cargoDto.getType(), cargoDto.getWeight()));
-    } //не знаю насчёт Id
+        Cargo newCargo = mapper.toEntity(cargoDto);
+        cargoRepository.save(newCargo);
+    }
 
     public void deleteById(Integer id){
-        cargoRepository.deleteById(id);
+        Optional<Cargo> cargoOptional = cargoRepository.findById(id);
+        if(cargoOptional.isPresent()){
+            cargoRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("Груз с id: " + id + " не найден");
+        }
+
     }
 }
